docs : landing page with translations and join page
Created a first version of landing page to guide the first visitors to join the waiting list and test the demonstration. The waiting list is created with cyptpad and embedded in the join page, but will be removed once the KYC is up and running. The demonstration site will be hosted on test.letspeppol.org and is subject to changes, therefore the printscreen will allways be outdated. Additionally, the 3 official Belgian languages have been added as options by using GPT-5 to do the translations.